Moved all the functions connected with player creation to a separate file called player.py
